Fixed bug where Walkers got related to same Survey Unit more than once
Adela's email:
6) I see that walker names are getting attached as children to Survey
Unit records as many times as there are rows. I sort of grasp why this
is happening, but is it going to cause performance issues? Could you
create a performance testing module for the Gridded survey, or simulate
walking some 70-100 units of 5x5 cells with random numbers within? On
days when we walk this many units, 80% of units have 0-1 shards in them.